Add pause time of 30 minutes.
The session will be paused after every 30 minutes for a total of 10
minutes before the user is allowed to login back. The pause time will
be universal (/var/log/pause_time) because mostly the kids share the
screen and we want all the kids to wait before logging back on.